% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{cp}
\alias{cp}
\title{Cambio porcentual}
\usage{
cp(ayer, hoy)
}
\arguments{
\item{ayer}{Valor inicial}

\item{hoy}{Valor final}
}
\description{
Funcion sencilla para calcular el cambio porcentual entre dos periodos
}
