% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getONC.R
\name{getONC}
\alias{getONC}
\title{getONC}
\usage{
getONC(nivel, valores, nom_periodo, modalidades = FALSE)
}
\arguments{
\item{nivel}{Valor string puede ser "estatal" o "municipal" para consultar los datos estatales o municipales}

\item{valores}{Valor string para definir los valores consultados de la base, opciones: "ci_tasa" para tasa de carpetas de investigacion, "ci_abs" para numero absoluto de carpetas, "vic_tasa" para tasa de victimas y "vic_abs" para numero absoluto de victimas.
Tambien se puede definir como un vector que contenga mas de uno de los valores}

\item{nom_periodo}{Valor string para definir el periodo temporal de los datos, opciones: "mensual", "bimestral", "trimestral", "cuatrimestral", "semestral" o "anual"}

\item{modalidades}{Argumento para incluir datos sobre modalidad del delito y caracteristicas de las victimas}
}
\value{

}
\description{
Consultar los datos de incidencia delictiva del SESNSP limpiados por el Observatorio Nacional Ciudadano
}
\examples{
getONC("estatal", c("ci_tasa", "vic_abs"), "mensual", modalidades = FALSE)

getONC("estatal", c("ci_tasa", "vic_abs"), "semestral", modalidades = TRUE)

getONC("municipal", "ci_abs", "anual", modalidades = FALSE)
}
